# src/docs/openapi2/swagger.yaml
definitions:
...
  consts.ResourceName:
    enum:
    - system
    - audit
    - container
    - container_version
    - dataset
    - dataset_version
    - project
    - label
    - user
    - role
    - permission
    - task
    - trace
    - injection
    - execution
    type: string
    x-enum-comments:
      ResourceAudit: audit resource
      ResourceContainer: container resource
      ResourceContainerVersion: container version resource
      ResourceDataset: dataset resource
      ResourceDatasetVersion: dataset version resource
      ResourceExecution: execution resource
      ResourceInjection: fault injection resource
      ResourceLabel: label resource
      ResourcePermission: permission resource
      ResourceProject: project resource
      ResourceRole: role resource
      ResourceSystem: system resource
      ResourceTask: task resource
      ResourceTrace: trace resource
      ResourceUser: user resource
    x-enum-varnames:
    - ResourceSystem
    - ResourceAudit
    - ResourceContainer
    - ResourceContainerVersion
    - ResourceDataset
    - ResourceDatasetVersion
    - ResourceProject
    - ResourceLabel
    - ResourceUser
    - ResourceRole
    - ResourcePermission
    - ResourceTask
    - ResourceTrace
    - ResourceInjection
    - ResourceExecution
...
  dto.AuditLogDetailResp:
    properties:
      action:
        type: string
      created_at:
        type: string
      details:
        type: string
      duration:
        type: integer
      error_msg:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      resource:
        $ref: '#/definitions/consts.ResourceName'
      resource_id:
        type: integer
      state:
        type: string
      status:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  dto.AuditLogResp:
    properties:
      action:
        type: string
      created_at:
        type: string
      duration:
        type: integer
      id:
        type: integer
      ip_address:
        type: string
      resource:
        $ref: '#/definitions/consts.ResourceName'
      resource_id:
        type: integer
      state:
        type: string
      status:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  dto.GenericResponse-any:
    properties:
      code:
        description: Status code
        type: integer
      data:
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_AuditLogDetailResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.AuditLogDetailResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_ListNamespaceLockResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.ListNamespaceLockResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
  dto.GenericResponse-dto_ListResp-dto_AuditLogResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.ListResp-dto_AuditLogResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_QueuedTasksResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.QueuedTasksResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_SystemInfo:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.SystemInfo'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.ListNamespaceLockResp:
    properties:
      items:
        type: object
    type: object
  dto.ListResp-dto_AuditLogResp:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.AuditLogResp'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationInfo'
    type: object
...
  dto.PaginationInfo:
    properties:
      page:
        example: 1
        type: integer
      size:
        example: 20
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  dto.QueuedTasksResp:
    properties:
      delayed_tasks:
        items:
          $ref: '#/definitions/dto.TaskResp'
        type: array
      ready_tasks:
        items:
          $ref: '#/definitions/dto.TaskResp'
        type: array
    type: object
...
  dto.SystemInfo:
    properties:
      cpu_usage:
        type: number
      disk_usage:
        type: number
      load_average:
        type: string
      memory_usage:
        type: number
    type: object
...
  dto.TaskResp:
    properties:
      created_at:
        type: string
      cron_expr:
        type: string
      execute_time:
        type: integer
      group_id:
        type: string
      id:
        type: string
      immediate:
        type: boolean
      project_id:
        type: integer
      project_name:
        type: string
      state:
        type: string
      status:
        type: string
      trace_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: team@rcabench.com
    name: RCABench Team
  description: |-
    RCABench - A comprehensive root cause analysis benchmarking platform for microservices
    This API provides endpoints for managing datasets, algorithms, evaluations, and fault injections
    for root cause analysis in distributed systems and microservices architectures.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: RCABench API
  version: 1.0.1
paths:
...
  /system/audit:
    get:
      description: Get paginated list of audit logs with optional filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: size
        type: integer
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by resource ID
        in: query
        name: resource_id
        type: integer
      - description: Filter by state
        in: query
        name: state
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: integer
      - description: Filter from date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Filter to date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Audit logs retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_ListResp-dto_AuditLogResp'
        "400":
          description: Invalid request format/parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: List audit logs
      tags:
      - System
  /system/audit/{id}:
    get:
      description: Get a specific audit log entry by ID
      parameters:
      - description: Audit log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Audit log retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_AuditLogDetailResp'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: Audit log not found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Get audit log by ID
      tags:
      - System
...
  /system/monitor/info:
    get:
      description: Get basic system information and status
      produces:
      - application/json
      responses:
        "200":
          description: System info retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_SystemInfo'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Get system information
      tags:
      - System
...
  /system/monitor/namespace/locks:
    get:
      description: Retrieve the list of currently locked namespaces
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the list of locks
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_ListNamespaceLockResp'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: List namespace locks
      tags:
      - System
  /system/monitor/task/queue:
    post:
      description: List tasks in queue (ready and delayed)
      produces:
      - application/json
      responses:
        "200":
          description: Queued tasks retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_QueuedTasksResp'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: No queued tasks found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: List queued tasks
      tags:
      - Tasks
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"